store setup crete update delete with route

+++ src/controller/setup/storeSetupController.ts
@@ -1,8 +1,11 @@
 import { Request, Response, NextFunction } from "express";
 import { StoreSetupServie } from "../../services/storeSetupService";
+import { PrismaClient } from "@prisma/client";
+const prismaService = new PrismaClient();
 const storeSetupService = new StoreSetupServie();
+
 export class StoreSetupController {
-  async getAll(res: Response, next: NextFunction) {
+  async getAll(req: Request, res: Response, next: NextFunction) {
     try {
       const reslts = await storeSetupService.getAll(next);
       res.json({ success: true, reslts });
@@ -21,6 +24,16 @@ export class StoreSetupController {
 
   async update(req: Request, res: Response, next: NextFunction) {
     try {
+      const { id } = req.params;
+      const checkExists = await prismaService.store.findUnique({
+        where: {
+          id,
+        },
+      });
+
+      if (!checkExists) {
+        return res.status(404).json("Not Found");
+      }
       const reslts = await storeSetupService.update(req, res, next);
       res.json({ success: true, reslts });
     } catch (error) {
@@ -30,6 +43,16 @@ export class StoreSetupController {
 
   async deleteStore(req: Request, res: Response, next: NextFunction) {
     try {
+      const { id } = req.params;
+      const checkExists = await prismaService.store.findUnique({
+        where: {
+          id,
+        },
+      });
+
+      if (!checkExists) {
+        return res.status(404).json("Not Found");
+      }
       const reslts = await storeSetupService.deleteStore(req, res, next);
       res.json({ success: true, reslts });
     } catch (error) {
